Modularize viewBpToPx
